tesis/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
dataclasses==0.6
Django==1.11.9
django-bootstrap4==0.0.4
djongo==1.2.23
paho-mqtt==1.6.1
Pillow==9.2.0
ping3==4.0.3
psycopg2-binary==2.9.3
pymongo==3.2.2
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.2.4
